# apps/control/src/infra/boat/boat.openapi.json
{
  "components": {
    "parameters": {
...
      "SandboxId": {
        "description": "Public Sandbox id returned by create/list/get sandbox calls.",
        "in": "path",
        "name": "sandboxId",
        "required": true,
        "schema": {
          "pattern": "^bx_[23456789abcdefghjkmnpqrstuvwxyz]{8}$",
          "type": "string"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "examples": {
              "invalid": {
                "value": {
                  "code": "invalid_json",
                  "error": {
                    "code": "invalid_json",
                    "message": "Request body must be valid JSON.",
                    "status": 400
                  },
                  "message": "Request body must be valid JSON.",
                  "ok": false,
                  "requestId": "req_01HX...",
                  "status": 400,
                  "type": "sandbox.error"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        },
        "description": "Invalid request body or parameters."
      },
...
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "examples": [
              "provider_not_configured"
            ],
            "type": "string"
          },
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              },
              "status": {
                "type": "integer"
              }
            },
            "required": [
              "code",
              "message",
              "status"
            ],
            "type": "object"
          },
          "message": {
            "examples": [
              "Prompting is locked until Codex is configured on the Agents page."
            ],
            "type": "string"
          },
          "ok": {
            "examples": [
              false
            ],
            "type": "boolean"
          },
          "requestId": {
            "examples": [
              "req_01HX..."
            ],
            "type": "string"
          },
          "status": {
            "examples": [
              409
            ],
            "type": "integer"
          },
          "type": {
            "examples": [
              "sandbox.error"
            ],
            "type": "string"
          }
        },
        "required": [
          "ok",
          "type",
          "status",
          "code",
          "message",
          "error",
          "requestId"
        ],
        "type": "object"
      },
...
      "FileReadResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessBase"
          },
          {
            "properties": {
              "content": {
                "type": "string"
              },
              "encoding": {
                "enum": [
                  "utf8",
                  "base64"
                ],
                "type": "string"
              },
              "path": {
                "type": "string"
              },
              "size": {
                "type": "integer"
              },
              "success": {
                "type": "boolean"
              },
              "type": {
                "const": "file.read",
                "type": "string"
              }
            },
            "required": [
              "success",
              "path",
              "encoding",
              "size",
              "content"
            ],
            "type": "object"
          }
        ]
      },
      "FileWriteRequest": {
        "properties": {
          "content": {
            "type": "string"
          },
          "encoding": {
            "default": "utf8",
            "enum": [
              "utf8",
              "base64"
            ],
            "type": "string"
          },
          "path": {
            "description": "Absolute path, or path relative to the sandbox work directory (/home/user). The canonicalized path must resolve under /home/user or /tmp; anything else is rejected with a 400 invalid_path error.",
            "type": "string"
          }
        },
        "required": [
          "path",
          "content"
        ],
        "type": "object"
      },
      "FileWriteResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessBase"
          },
          {
            "properties": {
              "encoding": {
                "enum": [
                  "utf8",
                  "base64"
                ],
                "type": "string"
              },
              "path": {
                "type": "string"
              },
              "size": {
                "type": "integer"
              },
              "success": {
                "type": "boolean"
              },
              "type": {
                "const": "file.written",
                "type": "string"
              }
            },
            "required": [
              "success",
              "path",
              "encoding",
              "size"
            ],
            "type": "object"
          }
        ]
      },
...
      "SuccessBase": {
        "properties": {
          "ok": {
            "examples": [
              true
            ],
            "type": "boolean"
          },
          "type": {
            "description": "Stable success envelope discriminator added by v1.",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "type"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BoatBearerAuth": {
        "bearerFormat": "sandbox_api_key",
        "description": "Boat bearer token in the form `boat_...`. Service API keys authenticate sandbox operations.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Public JSON API for creating, operating, prompting, observing, and exposing sandboxes from backend services, CI jobs, hosted workers, and Boat automation products.\n\nThe v1 reference intentionally documents the developer integration surface only. Dashboard billing actions are not part of v1.\n",
    "title": "Boat Public API v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sandboxes/{sandboxId}/files": {
      "get": {
        "operationId": "readFile",
        "parameters": [
          {
            "$ref": "#/components/parameters/SandboxId"
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "description": "Absolute path, or path relative to the sandbox work directory (/home/user). The canonicalized path must resolve under /home/user or /tmp; anything else is rejected with a 400 invalid_path error.",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "encoding",
            "schema": {
              "default": "utf8",
              "enum": [
                "utf8",
                "base64"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileReadResponse"
                }
              }
            },
            "description": "File contents."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "summary": "Read a file from a sandbox",
        "tags": [
          "Boat"
        ]
      },
      "put": {
        "operationId": "writeFile",
        "parameters": [
          {
            "$ref": "#/components/parameters/SandboxId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileWriteResponse"
                }
              }
            },
            "description": "File written."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "summary": "Write a file in a sandbox",
        "tags": [
          "Boat"
        ]
      }
    },
...
  },
  "security": [
    {
      "BoatBearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://boat.dev/api/v1"
    }
  ],
...
}